LibWeb: Move the CSS Rust code into its own crate under CSS/Rust
The LibWeb Rust crate mixed CSS code (the CSS tokenizer, the selector
engine and the style value data) with the encoding detection used by
the HTML parser. Give the CSS side its own libweb_css_rust crate at
Libraries/LibWeb/CSS/Rust, following the per-subsystem crate layout
already used by HTML/Parser/Rust and ContentBlocker/Rust, so upcoming
style system work lands in a CSS-owned crate from the start.

The original crate keeps the encoding detection and the re-export that
links the HTML tokenizer crate's symbols into the archive. The
generated FFI header names and install locations are unchanged, so no
C++ includes change. The new crate follows the ContentBlocker pattern
for the Rust allocator: it defines one only in static builds, behind
an allocator feature, and otherwise relies on the allocator from
libweb_rust linked into the same library.
