+++ Libraries/LibWeb/CSS/Rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "libweb_rust"
+name = "libweb_css_rust"
 version = "0.1.0"
 edition = "2024"
 
@@ -8,12 +8,13 @@ crate-type = ["staticlib"]
 
 # After changing dependencies, regenerate the Flatpak sources:
 #   python3 Meta/CMake/flatpak/generate-cargo-sources.py
-[dependencies]
-libweb_html_tokenizer = { path = "../HTML/Parser/Rust" }
-chardetng = "1.0.0"
-
 [build-dependencies]
 cbindgen = "0.29"
+serde_json = { version = "1", features = ["preserve_order"] }
+
+[features]
+default = []
+allocator = []
 
 [lints]
 workspace = true
